+++ toolkit/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "2.3.5"
+__version__ = "2.4.0"
 __author__ = "anonymous"